requirements.txt
streamlit
openai-whisper
moviepy
transformers
torch